chore (code-refactoring): updated modules
